remove entry point changes
