--- src/constants/horizontalPanel.ts
@@ -1,11 +1,0 @@
-export enum InfoSection {
-  Position = 'Position',
-  Orders = 'Orders',
-  Fills = 'Fills',
-  Payments = 'Payments',
-}
-
-export enum PanelView {
-  AllMarkets = 'AllMarkets',
-  CurrentMarket = 'CurrentMarket',
-}
