+++ src/pages/trade/Trade.tsx
@@ -3,7 +3,6 @@ import { useRef, useState } from 'react';
 import { useSelector } from 'react-redux';
 import styled, { AnyStyledComponent, css } from 'styled-components';
 
-import { PanelView, InfoSection } from '@/constants/horizontalPanel';
 import { TradeLayouts } from '@/constants/layout';
 
 import {
@@ -21,7 +20,6 @@ import { AccountInfo } from '@/views/AccountInfo';
 import { TradeBox } from '@/views/TradeBox';
 
 import { calculateCanAccountTrade } from '@/state/accountCalculators';
-import { getDefaultToAllMarketsInPositionsOrdersFills } from '@/state/configsSelectors';
 import { getSelectedTradeLayout } from '@/state/layoutSelectors';
 
 import { HorizontalPanel } from './HorizontalPanel';
@@ -42,11 +40,6 @@ const TradePage = () => {
   const canAccountTrade = useSelector(calculateCanAccountTrade);
 
   const [isHorizontalPanelOpen, setIsHorizontalPanelOpen] = useState(true);
-  const allMarkets = useSelector(getDefaultToAllMarketsInPositionsOrdersFills);
-  const [view, setView] = useState<PanelView>(
-    allMarkets ? PanelView.AllMarkets : PanelView.CurrentMarket
-  );
-  const [tab, setTab] = useState<InfoSection>(InfoSection.Position);
 
   usePageTitlePriceUpdates();
   useTradeFormInputs();
@@ -61,7 +54,7 @@ const TradePage = () => {
         </DetachedSection>
 
         <DetachedSection>
-          <HorizontalPanel tab={tab} view={view} setTab={setTab} setView={setView} />
+          <HorizontalPanel />
         </DetachedSection>
 
         <DetachedSection>
@@ -83,7 +76,7 @@ const TradePage = () => {
 
       <Styled.SideSection gridArea="Side">
         <AccountInfo />
-        <TradeBox setTab={setTab} setView={setView} />
+        <TradeBox />
       </Styled.SideSection>
 
       <Styled.GridSection gridArea="Vertical">
@@ -95,14 +88,7 @@ const TradePage = () => {
       </Styled.GridSection>
 
       <Styled.GridSection gridArea="Horizontal">
-        <HorizontalPanel
-          tab={tab}
-          view={view}
-          setTab={setTab}
-          setView={setView}
-          isOpen={isHorizontalPanelOpen}
-          setIsOpen={setIsHorizontalPanelOpen}
-        />
+        <HorizontalPanel isOpen={isHorizontalPanelOpen} setIsOpen={setIsHorizontalPanelOpen} />
       </Styled.GridSection>
     </Styled.TradeLayout>
   );
