+++ src/views/TradeBox.tsx
@@ -1,10 +1,7 @@
-import { Dispatch, SetStateAction } from 'react';
-
 import { useDispatch, useSelector } from 'react-redux';
 import styled, { type AnyStyledComponent } from 'styled-components';
 
 import { TradeBoxDialogTypes } from '@/constants/dialogs';
-import { PanelView, InfoSection } from '@/constants/horizontalPanel';
 import { STRING_KEYS } from '@/constants/localization';
 
 import { useStringGetter } from '@/hooks';
@@ -22,12 +19,7 @@ import abacusStateManager from '@/lib/abacus';
 
 import { TradeBoxOrderView } from './TradeBoxOrderView';
 
-type ElementProps = {
-  setTab: Dispatch<SetStateAction<InfoSection>>;
-  setView: Dispatch<SetStateAction<PanelView>>;
-};
-
-export const TradeBox = ({ setTab, setView }: ElementProps) => {
+export const TradeBox = () => {
   const dispatch = useDispatch();
   const stringGetter = useStringGetter();
 
@@ -55,7 +47,7 @@ export const TradeBox = ({ setTab, setView }: ElementProps) => {
 
   return (
     <Styled.TradeBox>
-      <TradeBoxOrderView setTab={setTab} setView={setView} />
+      <TradeBoxOrderView />
 
       <Styled.Dialog
         isOpen={!!activeDialog}
