+++ src/views/TradeBoxOrderView.tsx
@@ -1,11 +1,10 @@
-import { Dispatch, SetStateAction, useCallback } from 'react';
+import { useCallback } from 'react';
 
 import { shallowEqual, useSelector } from 'react-redux';
 import { createSelector } from 'reselect';
 import styled, { AnyStyledComponent } from 'styled-components';
 
 import { TradeInputField } from '@/constants/abacus';
-import { PanelView, InfoSection } from '@/constants/horizontalPanel';
 import { STRING_KEYS, StringKey } from '@/constants/localization';
 import { TradeTypes } from '@/constants/trade';
 
@@ -63,12 +62,7 @@ const useTradeTypeOptions = () => {
   };
 };
 
-type ElementProps = {
-  setTab: Dispatch<SetStateAction<InfoSection>>;
-  setView: Dispatch<SetStateAction<PanelView>>;
-};
-
-export const TradeBoxOrderView = ({ setTab, setView }: ElementProps) => {
+export const TradeBoxOrderView = () => {
   const onTradeTypeChange = useCallback((tradeType?: TradeTypes) => {
     if (tradeType) {
       abacusStateManager.clearTradeInputValues();
@@ -86,7 +80,7 @@ export const TradeBoxOrderView = ({ setTab, setView }: ElementProps) => {
       onValueChange={onTradeTypeChange}
       sharedContent={
         <Styled.Container>
-          <TradeForm setTab={setTab} setView={setView} />
+          <TradeForm />
         </Styled.Container>
       }
       fullWidthTabs
